[quant][pt2e] Rename _pt2e to pt2e
Summary:
att

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
