+++ torch/ao/quantization/pt2e/quantizer/qnnpack_quantizer.py
@@ -11,9 +11,9 @@ import torch
 import torch._dynamo as torchdynamo
 import torch.nn.functional as F
 
-from torch.ao.quantization._pt2e.graph_utils import find_sequential_partitions
+from torch.ao.quantization.pt2e.graph_utils import find_sequential_partitions
 
-from torch.ao.quantization._pt2e.quantizer.utils import (
+from torch.ao.quantization.pt2e.quantizer.utils import (
     _annotate_input_qspec_map,
     _annotate_output_qspec,
     _is_sym_size_node,
