+++ torch/ao/quantization/pt2e/quantizer/x86_inductor_quantizer.py
@@ -11,7 +11,7 @@ from .quantizer import (
     Quantizer,
     QuantizationAnnotation,
 )
-from torch.ao.quantization._pt2e.quantizer.utils import (
+from torch.ao.quantization.pt2e.quantizer.utils import (
     get_input_act_qspec,
     get_output_act_qspec,
     get_weight_qspec,
